[TRL-283] refactor: DayQueryRepository 구조 개선
